# quizverse/backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "QuizVerse API",
        "description": "API dokumentacji dla aplikacji QuizVerse",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/questions/is-correct/{id}": {
            "post": {
                "tags": [
                    "Answers"
                ],
                "summary": "Sprawdź odpowiedź",
                "description": "Sprawdza czy odpowiedź jest poprawna",
                "operationId": "556e1a767b159f587ff8c41081da958b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID odpowiedzi",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status odpowiedzi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "answer_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "question_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "is_correct": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Odpowiedź nie znaleziona"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/questions/{questionId}/answers": {
            "post": {
                "tags": [
                    "Answers"
                ],
                "summary": "Dodaj odpowiedź",
                "description": "Dodaje nową odpowiedź do pytania (tylko dla właściciela quizu)",
                "operationId": "8960a6a3e4ac45e90bb7fc102dee2003",
                "parameters": [
                    {
                        "name": "questionId",
                        "in": "path",
                        "description": "ID pytania",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "text",
                                    "is_correct"
                                ],
                                "properties": {
                                    "text": {
                                        "type": "string",
                                        "example": "Warszawa"
                                    },
                                    "is_correct": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "image_path": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Odpowiedź została dodana",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Answer added successfully"
                                        },
                                        "answer": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "question_id": {
                                                    "type": "integer"
                                                },
                                                "text": {
                                                    "type": "string"
                                                },
                                                "is_correct": {
                                                    "type": "boolean"
                                                },
                                                "image_path": {
                                                    "type": "string",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Brak uprawnień"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Lista kategorii",
                "description": "Pobiera listę wszystkich kategorii",
                "operationId": "b1b247b2b491b0491c3c728dc33f23f0",
                "responses": {
                    "200": {
                        "description": "Lista kategorii",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "categories": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Historia"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/categories/select": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Kategorie dla selecta",
                "description": "Pobiera kategorie w formacie id => name dla selectów",
                "operationId": "d7ecc6fca854022002170c46e8b89d88",
                "responses": {
                    "200": {
                        "description": "Kategorie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "categories": {
                                            "type": "object",
                                            "example": {
                                                "1": "Historia",
                                                "2": "Geografia"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}